+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
+- `_model_function` method refactored
+- `_model_function` method renamed to `model_function`
 - Dependencies structure modified
 ## [0.4] - 2026-05-18
 ### Added